+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.4.0
+- New build task to trigger build+install in one step
+- Device status bar now pops up a menu with device related commands
+- New command to open Device logs
+
 ## 0.3.0
 - All commands reworked to be functional from the command palette
 - Commands to add and remove remotes